backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI APIs
google-generativeai==0.3.1
google-cloud-speech==2.21.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.1

# Utilities
pydantic==2.5.0

# For Vercel deployment (optional but recommended)
mangum==0.17.0